Renamed triple2grid_wrapper to triple_to_grid_wrapper
